backend/requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==6.0.1
markdown2==2.5.4
latex2mathml==3.78.0
chromadb==0.6.3
numpy<2.0
PyMuPDF==1.26.3
python-dotenv==1.0.0
pydantic==2.11.7
requests==2.31.0
aiohttp==3.9.1
openai==1.102.0
pytest==7.4.3
pytest-flask==1.3.0
grobid_tei_xml==0.1.6